chore: update to provide App class in core project
