--- src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Base/App.xaml
@@ -1,45 +1,0 @@
-<!--#if (useCsharpMarkup)-->
-<Application x:Class="MyExtensionsApp.App"
-			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-<!--#else-->
-<Application x:Class="MyExtensionsApp.App"
-			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:wasm="http://platform.uno/wasm"
-			 xmlns:local="using:MyExtensionsApp"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 mc:Ignorable="wasm">
-
-	<Application.Resources>
-		<ResourceDictionary>
-			<ResourceDictionary.MergedDictionaries>
-				<!-- Load WinUI resources -->
-				<XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-<!--#if (useMaterial)-->
-				<MaterialColors xmlns="using:Uno.Material" />
-								<!-- Temporarily removing due to issue in Uno.Material-->
-								<!--OverrideSource="ms-appx:///MyExtensionsApp/Styles/ColorPaletteOverride.xaml" />-->
-				<MaterialFonts xmlns="using:Uno.Material" />
-								<!--OverrideSource="ms-appx:///MyExtensionsApp/Styles/MaterialFontsOverride.xaml" />-->
-				<MaterialResources xmlns="using:Uno.Material" />
-<!--#elif (useCupertino)-->
-				<CupertinoColors xmlns="using:Uno.Cupertino" />
-				<CupertinoFonts xmlns="using:Uno.Cupertino" />
-				<CupertinoResources xmlns="using:Uno.Cupertino" />
-<!--#endif-->
-<!--#if (useRecommendedAppTemplate)-->
-
-				<!-- Load Uno.UI.Toolkit resources -->
-				<ToolkitResources xmlns="using:Uno.Toolkit.UI" />
-<!--#endif-->
-<!--#if (useMaterial)-->
-				<MaterialToolkitResources xmlns="using:Uno.Toolkit.UI.Material" />
-<!--#endif-->
-			</ResourceDictionary.MergedDictionaries>
-			<!-- Add resources here -->
-		</ResourceDictionary>
-	</Application.Resources>
-
-<!--#endif-->
-</Application>
