+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Base/AppHead.xaml
@@ -1,0 +1,24 @@
+<!--#if (useCsharpMarkup)-->
+<local:App x:Class="MyExtensionsApp.AppHead"
+			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+			 xmlns:local="using:MyExtensionsApp">
+<!--#else-->
+<local:App x:Class="MyExtensionsApp.AppHead"
+			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+			 xmlns:wasm="http://platform.uno/wasm"
+			 xmlns:local="using:MyExtensionsApp"
+			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+			 mc:Ignorable="wasm">
+
+	<local:App.Resources>
+		<ResourceDictionary>
+			<ResourceDictionary.MergedDictionaries>
+				<ResourceDictionary Source="ms-appx:///MyExtensionsApp/AppResources.xaml" />
+			</ResourceDictionary.MergedDictionaries>
+		</ResourceDictionary>
+	</local:App.Resources>
+
+<!--#endif-->
+</local:App>
