+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Base/AppHead.xaml.cs
@@ -1,0 +1,117 @@
+//+:cnd:noEmit
+#if (useBlankAppTemplate)
+using System;
+using Microsoft.Extensions.Logging;
+using Windows.ApplicationModel;
+using Windows.ApplicationModel.Activation;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
+using Microsoft.UI.Xaml.Navigation;
+
+#endif
+namespace MyExtensionsApp;
+
+public sealed partial class AppHead : App
+{
+#if (useBlankAppTemplate)
+	static AppHead() =>
+		InitializeLogging();
+
+#endif
+	/// <summary>
+	/// Initializes the singleton application object. This is the first line of authored code
+	/// executed, and as such is the logical equivalent of main() or WinMain().
+	/// </summary>
+	public AppHead()
+	{
+		this.InitializeComponent();
+	}
+#if useCsharpMarkup
+
+	/// <summary>
+	/// Invoked when the application is launched normally by the end user.  Other entry points
+	/// will be used such as when the application is launched to open a specific file.
+	/// </summary>
+	/// <param name="args">Details about the launch request and process.</param>
+	protected override void OnLaunched(LaunchActivatedEventArgs args)
+	{
+		Resources.Build(r => r.Merged(
+			new AppResources()));
+
+		base.OnLaunched(args);
+	}
+#endif
+#if (useBlankAppTemplate)
+
+	/// <summary>
+	/// Configures global Uno Platform logging
+	/// </summary>
+	private static void InitializeLogging()
+	{
+//-:cnd:noEmit
+#if DEBUG
+		// Logging is disabled by default for release builds, as it incurs a significant
+		// initialization cost from Microsoft.Extensions.Logging setup. If startup performance
+		// is a concern for your application, keep this disabled. If you're running on the web or
+		// desktop targets, you can use URL or command line parameters to enable it.
+		//
+		// For more performance documentation: https://platform.uno/docs/articles/Uno-UI-Performance.html
+
+		var factory = LoggerFactory.Create(builder =>
+		{
+#if __WASM__
+			builder.AddProvider(new global::Uno.Extensions.Logging.WebAssembly.WebAssemblyConsoleLoggerProvider());
+#elif __IOS__ && !__MACCATALYST__
+			builder.AddProvider(new global::Uno.Extensions.Logging.OSLogLoggerProvider());
+#elif NETFX_CORE
+			builder.AddDebug();
+#else
+			builder.AddConsole();
+#endif
+
+			// Exclude logs below this level
+			builder.SetMinimumLevel(LogLevel.Information);
+
+			// Default filters for Uno Platform namespaces
+			builder.AddFilter("Uno", LogLevel.Warning);
+			builder.AddFilter("Windows", LogLevel.Warning);
+			builder.AddFilter("Microsoft", LogLevel.Warning);
+
+			// Generic Xaml events
+			// builder.AddFilter("Microsoft.UI.Xaml", LogLevel.Debug );
+			// builder.AddFilter("Microsoft.UI.Xaml.VisualStateGroup", LogLevel.Debug );
+			// builder.AddFilter("Microsoft.UI.Xaml.StateTriggerBase", LogLevel.Debug );
+			// builder.AddFilter("Microsoft.UI.Xaml.UIElement", LogLevel.Debug );
+			// builder.AddFilter("Microsoft.UI.Xaml.FrameworkElement", LogLevel.Trace );
+
+			// Layouter specific messages
+			// builder.AddFilter("Microsoft.UI.Xaml.Controls", LogLevel.Debug );
+			// builder.AddFilter("Microsoft.UI.Xaml.Controls.Layouter", LogLevel.Debug );
+			// builder.AddFilter("Microsoft.UI.Xaml.Controls.Panel", LogLevel.Debug );
+
+			// builder.AddFilter("Windows.Storage", LogLevel.Debug );
+
+			// Binding related messages
+			// builder.AddFilter("Microsoft.UI.Xaml.Data", LogLevel.Debug );
+			// builder.AddFilter("Microsoft.UI.Xaml.Data", LogLevel.Debug );
+
+			// Binder memory references tracking
+			// builder.AddFilter("Uno.UI.DataBinding.BinderReferenceHolder", LogLevel.Debug );
+
+			// RemoteControl and HotReload related
+			// builder.AddFilter("Uno.UI.RemoteControl", LogLevel.Information);
+
+			// Debug JS interop
+			// builder.AddFilter("Uno.Foundation.WebAssemblyRuntime", LogLevel.Debug );
+		});
+
+		global::Uno.Extensions.LogExtensionPoint.AmbientLoggerFactory = factory;
+
+#if HAS_UNO
+		global::Uno.UI.Adapter.Microsoft.Extensions.Logging.LoggingAdapter.Initialize();
+#endif
+#endif
+//+:cnd:noEmit
+	}
+#endif
+}
