+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Base/base.props
@@ -1,14 +1,14 @@
 <Project>
 	<ItemGroup>
-		<None Include="$(MSBuildThisFileDirectory)App.xaml" />
-		<ApplicationDefinition Include="$(MSBuildThisFileDirectory)App.xaml"
+		<None Include="$(MSBuildThisFileDirectory)AppHead.xaml" />
+		<ApplicationDefinition Include="$(MSBuildThisFileDirectory)AppHead.xaml"
 								SubType="Designer"
 								XamlRuntime="WinUI"
 								Generator="MSBuild:Compile"
-								Link="App.xaml" />
-		<Compile Include="$(MSBuildThisFileDirectory)App.xaml.cs"
+								Link="AppHead.xaml" />
+		<Compile Include="$(MSBuildThisFileDirectory)AppHead.xaml.cs"
 				XamlRuntime="WinUI"
-				DependentUpon="App.xaml"
-				Link="App.xaml.cs" />
+				DependentUpon="AppHead.xaml"
+				Link="AppHead.xaml.cs" />
 	</ItemGroup>
 </Project>
