+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp/App.blank.cs
@@ -1,46 +1,22 @@
 //-:cnd:noEmit
 namespace MyExtensionsApp;
 
-public class AppStart
+public class App : Application
 {
-	public static Window? Window { get; private set; }
+	public static Window? _window;
 
-	public static void OnLaunched(Application app, LaunchActivatedEventArgs args)
+	public void OnLaunched(LaunchActivatedEventArgs args)
 	{
-//+:cnd:noEmit
-#if useCsharpMarkup
-		// Load WinUI Resources
-		app.Resources(r => r.Merged(
-			new XamlControlsResources()));
-
-#if useMaterial
-		// Load Material Resources
-		app.UseMaterial(
-			colorOverride: new Styles.ColorPaletteOverride(),
-			fontOverride: new Styles.MaterialFontsOverride());
-
-		// Load Uno.UI.Toolkit Resources
-		app.Resources(r => r.Merged(
-			new ToolkitResources(),
-			new MaterialToolkitResources()));
-#else
-		// Load Uno.UI.Toolkit Resources
-		app.Resources(r => r.Merged(
-			new ToolkitResources()));
-#endif
-
-#endif
-//-:cnd:noEmit
 #if NET6_0_OR_GREATER && WINDOWS && !HAS_UNO
-		Window = new Window();
-		Window.Activate();
+		_window = new Window();
+		_window.Activate();
 #else
-		Window = Microsoft.UI.Xaml.Window.Current;
+		_window = Microsoft.UI.Xaml.Window.Current;
 #endif
 
 		// Do not repeat app initialization when the Window already has content,
 		// just ensure that the window is active
-		if (Window.Content is not Frame rootFrame)
+		if (_window.Content is not Frame rootFrame)
 		{
 			// Create a Frame to act as the navigation context and navigate to the first page
 			rootFrame = new Frame();
@@ -53,7 +29,7 @@ public class AppStart
 			}
 
 			// Place the frame in the current Window
-			Window.Content = rootFrame;
+			_window.Content = rootFrame;
 		}
 
 #if !(NETSTANDARD && WINDOWS)
@@ -68,7 +44,7 @@ public class AppStart
 				rootFrame.Navigate(typeof(MainPage), args.Arguments);
 			}
 			// Ensure the current window is active
-			Window.Activate();
+			_window.Activate();
 		}
 	}
